+++ routes/index.routes.js
@@ -1,7 +1,9 @@
 const express = require('express')
 const router = express.Router()
 const passport = require('passport')
-const userController = require('../controllers/user.controller')
+const userController = require('../controllers/user.controllers/user.controller')
+const userEditController = require('../controllers/user.controllers/user.edit.controller')
+const userSettingsController = require('../controllers/user.controllers/user.settings.controller')
 const orgController = require('../controllers/org.controller')
 const secure = require('../middlewares/secure.middleware')
 const upload = require('../configs/storage.config')
@@ -21,8 +23,6 @@ router.post(
 )
 router.get('/login', secure.isNotAuthenticated, userController.login)
 router.post('/login', secure.isNotAuthenticated, userController.doLogin)
-
-// Activated user routes ----------------------------------------------------------
 router.get(
     '/user/profile',
     secure.checkRoles('USER'),
@@ -35,60 +35,6 @@ router.get(
     secure.isNotAuthenticated,
     userController.activate
 )
-router.get(
-    '/activate-email/:token',
-    secure.isNotAuthenticated,
-    userController.activateNewEmail
-)
-router.get(
-    '/change-inaction/:token',
-    secure.isNotAuthenticated,
-    userController.activateInAction
-)
-router.get(
-    '/user/edit-profile',
-    secure.checkRoles('USER'),
-    secure.isAuthenticated,
-    userController.editProfile
-)
-router.post(
-    '/edit-profile',
-    secure.isAuthenticated,
-    upload.single('profilePicture'),
-    userController.doEditProfile
-)
-router.get(
-    '/user/settings',
-    secure.checkRoles('USER'),
-    secure.isAuthenticated,
-    userController.settings
-)
-router.post(
-    '/settings-e',
-    secure.isAuthenticated,
-    userController.doSettingsEmail
-)
-router.post(
-    '/settings-p',
-    secure.isAuthenticated,
-    userController.doSettingsPassword
-)
-router.post(
-    '/do-the-action',
-    secure.isNotAuthenticated,
-    userController.doTheAction
-)
-router.post(
-    '/settings-delete-account',
-    secure.isAuthenticated,
-    userController.doDelete
-)
-router.post(
-    '/settings-b',
-    secure.isAuthenticated,
-    userController.doSettingsBank
-)
-
 // Google auth
 router.get(
     '/authenticate/google',
@@ -103,6 +49,62 @@ router.get(
 )
 router.get('/auth/facebook/callback', userController.doLoginFacebook)
 
+// Edit user ---------------------------------------------------------
+router.get(
+    '/user/edit-profile',
+    secure.checkRoles('USER'),
+    secure.isAuthenticated,
+    userEditController.editProfile
+)
+router.post(
+    '/edit-profile',
+    secure.isAuthenticated,
+    upload.single('profilePicture'),
+    userEditController.doEditProfile
+)
+// Settigs user --------------------------------------------------------
+router.get(
+    '/user/settings',
+    secure.checkRoles('USER'),
+    secure.isAuthenticated,
+    userSettingsController.settings
+)
+router.post(
+    '/settings-e',
+    secure.isAuthenticated,
+    userSettingsController.doSettingsEmail
+)
+router.get(
+    '/activate-email/:token',
+    secure.isNotAuthenticated,
+    userSettingsController.activateNewEmail
+)
+router.post(
+    '/settings-p',
+    secure.isAuthenticated,
+    userSettingsController.doSettingsPassword
+)
+router.get(
+    '/change-inaction/:token',
+    secure.isNotAuthenticated,
+    userSettingsController.activateInAction
+)
+router.post(
+    '/do-the-action',
+    secure.isNotAuthenticated,
+    userSettingsController.doTheAction
+)
+router.post(
+    '/settings-b',
+    secure.isAuthenticated,
+    userSettingsController.doSettingsBank
+)
+router.post(
+    '/settings-delete-account',
+    secure.isAuthenticated,
+    userSettingsController.doDelete
+)
+
 // Org routes =======================================================================
 router.get('/org/register', secure.isNotAuthenticated, orgController.register)
 router.post(
@@ -111,8 +113,6 @@ router.post(
     upload.single('profilePicture'),
     orgController.doRegister
 )
-
-// Active org routes -----------------------------------------------------------------
 router.get(
     '/org/profile',
     secure.checkRoles('ORG'),
